LPS-115161 Refactor code and constants
